service implement

+++ src/app/usecases/mixin/get_directories.go
@@ -4,10 +4,10 @@ type GetDirectoryCaseRes struct {
 }
 
 type GetDirectoriesCaseRes struct {
-	PageIndex    int                   `json:"pageIndex" example:"1" format:"int"`
-	PageSize     int                   `json:"pageSize" example:"10" format:"int"`
-	TotalCount   int                   `json:"totalCount" example:"50" format:"int"`
-	TotalPages   int                   `json:"totalPages" example:"5" format:"int"`
+	PageIndex    int64                 `json:"pageIndex" example:"1" format:"int"`
+	PageSize     int64                 `json:"pageSize" example:"10" format:"int"`
+	TotalCount   int64                 `json:"totalCount" example:"50" format:"int"`
+	TotalPages   int64                 `json:"totalPages" example:"5" format:"int"`
 	Items        []GetDirectoryCaseRes `json:"items" format:"[]GetDirectoryCaseRes"`
 	HasPrevPages bool                  `json:"hasPrevPages" example:"true" format:"bool"`
 	HasNextPages bool                  `json:"hasNextPages" example:"true" format:"bool"`
